+++ container-engine-lib/lib/backend_interface/objects/service/service.go
@@ -24,7 +24,7 @@ type Service struct {
 
 	// These will only be non-nil if all of the following are true:
 	//  - The backend is Docker
-	//  - The service is in UserServiceStatus_Running state
+	//  - The service is in UserServiceStatus_Activated state
 	maybePublicIp    net.IP                         // The ip exposed in the host machine. Will be nil if the service doesn't declare any private ports
 	maybePublicPorts map[string]*port_spec.PortSpec //Mapping of port-used-by-service -> port-on-the-host-machine where the user can make requests to the port to access the port. If a used port doesn't have a host port bound, then the value will be nil.
 }